requirements.txt
opencv_python==4.13.0.92
pyserial==3.5
